backend/atsresume/requirements.txt
Django==3.1.12
djangorestframework==3.12.4
pymongo==3.11.4
PyJWT==2.8.0
djongo==1.3.7
python-dotenv==0.21.1
djangorestframework-simplejwt==5.3.0